refactor(workspace): share reconciled message types (P5.4c)
Third drifted-type slice — message contracts reconcile against backend
truth into @babyjamjam/shared/types/message:
- TemplateVariable regains the backend fields both apps had stripped
  (placeholder, optionType, options, dataSource, min, max); the
  GET /message-templates contract corrected from invented pagination to
  the actual array response
- SendMessageDeliverySmsRequest/Response shared with clientId, resolved
  vs request SMS types, provider msgType, optional request.senderPhone
- sender-approval + system-admin approval view contracts deduplicated
- RawMessageRecord/MessageRecord + create/update requests shared;
  UI-derived history/thread view models deliberately stay app-local

Gates: both type-checks clean, 632 + 245 tests green, lint counts
unchanged, both production builds exit 0.
