+++ packages/shared/package.json
@@ -10,6 +10,7 @@
         ".": "./src/index.ts",
         "./types/alimtalk": "./src/types/alimtalk.ts",
         "./types/eformsign": "./src/types/eformsign.ts",
+        "./types/message": "./src/types/message.ts",
         "./utils": "./src/utils/index.ts",
         "./eslint-plugin": "./eslint-plugin/index.mjs"
     },
